==> src/SDi2024/16_data_validation.R <==
# title: SoilData Integration
# subtitle: 
# author: Alessandro Samuel-Rosa
# date: 2025
# licence: MIT
# summary: This script performs data validation and cleaning on the integrated soil dataset. It 
#          starts by removing obsolete columns and standardizing column names. It then applies a 
#          series of validation rules to ensure data integrity, such as inferring missing sample 
#          types, cleaning coordinate datums, and checking for plausible ranges in soil properties 
#          like bulk density (dsi) and pH based on soil classification and horizon type. Values 
#          outside expected ranges are set to NA.

rm(list = ls())

# Install and load required packages
if (!requireNamespace("data.table")) {
  install.packages("data.table")
}
if (!requireNamespace("sf")) {
  install.packages("sf")
}
if (!requireNamespace("geobr")) {
  install.packages("geobr")
}

# Source helper functions
source("src/00_helper_functions.R")

# Read the Brazilian state boundaries
# Check if the file already exists to avoid re-downloading
if (!file.exists("data/brazil_states.geojson")) {
  brazil <- geobr::read_state(simplified = FALSE)
  # Save the data to a file for future use
  sf::st_write(brazil, "data/brazil_states.geojson")
} else {
  brazil <- sf::st_read("data/brazil_states.geojson")
}

# Read the Brazilian Soil Dataset v2023
# Check if file "data/00_brazilian_soil_dataset_2023.txt" exists. If not, read the Brazilian 
# Soil Dataset v2023 from the SoilData repository using the 'dataverse' package. Next, write the 
# file to 'data/00_brazilian_soil_dataset_2023.txt'. The dataset is available at
# https://doi.org/10.60502/SoilData/TUI25K. If the file already exists, read it using the
# 'data.table' package.
file_path <- "data/00_brazilian_soil_dataset_2023.txt"
if (!file.exists(file_path)) {
  br_soil2023 <- dataverse::get_dataframe_by_name("brazilian-soil-dataset-2023.txt",
    server = "https://soildata.mapbiomas.org/dataverse/soildata",
    dataset = "10.60502/SoilData/TUI25K", .f = data.table::fread
  )
  data.table::fwrite(br_soil2023, file_path, dec = ".", sep = ";")
} else {
  br_soil2023 <- data.table::fread(file_path, dec = ".", sep = ";")
}

# Read SoilData data processed in the previous step
soildata <- data.table::fread("data/15_soildata.txt", sep = "\t", na.strings = c("NA", ""))

# all columns in br_soil2023 are in soildata?
all(colnames(br_soil2023) %in% colnames(soildata))

# which columns in soildata are not in br_soil2023?
setdiff(colnames(soildata), colnames(br_soil2023))

summary_soildata(soildata)
# Layers: 57077
# Events: 16824
# Georeferenced events: 14334
# Datasets: 255

# Drop unwanted columns from soildata:
# acidez_caoac2ph7_naoh aluminio_kcl_naoh aluminio_saturacao_calc area_conhecimento autor_nome bases_saturacao_calc bases_soma_calc calcio_kcl_eaa campanha_id_febr ctce_soma_calc data_coleta_dia data_coleta_mes dataset_versao erosao_presenca erosao_tipo fosforo_mehlich1_eam magnesio_kcl_eaa palavras_chave potassio_mehlich1_eeac publicacao_data relevo_exposicao relevo_local relevo_posicao observacao_data
soildata <- soildata[, -c(
  "acidez_caoac2ph7_naoh", "aluminio_kcl_naoh", "aluminio_saturacao_calc", "area_conhecimento",
  "autor_nome", "bases_saturacao_calc", "bases_soma_calc", "calcio_kcl_eaa", "campanha_id_febr",
  "ctce_soma_calc", "data_coleta_dia", "data_coleta_mes", "dataset_versao", "erosao_presenca",
  "erosao_presenca", "erosao_tipo", "fosforo_mehlich1_eam", "id", "magnesio_kcl_eaa",
  "palavras_chave", "potassio_mehlich1_eeac", "publicacao_data", "relevo_exposicao",
  "relevo_local", "relevo_posicao", "observacao_data"
)]

# Rename 'data_fonte' to 'ano_fonte' to avoid confusion with 'dados_fonte' used somewere else
setnames(soildata, "data_fonte", "ano_fonte")

paste(colnames(soildata), collapse = ", ")
# dataset_id, observacao_id, dataset_titulo, organizacao_nome, dataset_licenca, sisb_id, ibge_id, coord_x, coord_y, coord_precisao, coord_fonte, pais_id, estado_id, municipio_id, amostra_tipo, amostra_quanti, amostra_area, taxon_sibcs, taxon_st, taxon_wrb, camada_id, amostra_id, camada_nome, profund_sup, profund_inf, terrafina, argila, silte, areia, carbono, ctc, ph, dsi, ce, data_ano, id, ano_fonte, coord_datum, esqueleto
# View(soildata)

# dataset_id and dataset_titulo
# Should be consistent: the number of unique dataset_id should be equal to the number of unique
# dataset_titulo
if (length(unique(soildata$dataset_id)) > length(unique(soildata$dataset_titulo))) {
  print(soildata[is.na(dataset_titulo) | dataset_titulo == "", unique(dataset_id)])
}

# amostra_tipo
# If amostra_quanti == 1, then amostra_tipo should be "SIMPLES"
soildata[amostra_quanti == 1 & (is.na(amostra_tipo) | amostra_tipo == ""), amostra_tipo := "SIMPLES"]
# If amostra_quanti > 1 and amostra_tipo is NA, then amostra_tipo should be "COMPOSTA"
soildata[amostra_quanti > 1 & is.na(amostra_tipo), amostra_tipo := "COMPOSTA"]
# statistics
soildata[, .N, by = .(amostra_tipo)][order(amostra_tipo)]

# amostra_quanti
# If amostra_tipo is "SIMPLES" and amostra_quanti is NA, then amostra_quanti should be 1
soildata[amostra_tipo == "SIMPLES" & is.na(amostra_quanti), amostra_quanti := 1]
# statistics
soildata[, .N, by = .(amostra_quanti)][order(amostra_quanti)]
# cross table between amostra_tipo and amostra_quanti
table(soildata$amostra_tipo, soildata$amostra_quanti, useNA = "ifany")

# dataset_licenca
# if "CC-BY-4.0", then "CC BY 4.0"
soildata[dataset_licenca == "CC-BY-4.0", dataset_licenca := "CC BY 4.0"]
# if "CC BY-NC", then "CC BY-NC 4.0"
soildata[dataset_licenca == "CC BY-NC", dataset_licenca := "CC BY-NC 4.0"]
# statistics
soildata[, .N, by = .(dataset_licenca)][order(dataset_licenca)]

# pais_id
# if BR, then Brasil
soildata[pais_id == "BR", pais_id := "Brasil"]
# if NA of "", then Brasil
soildata[is.na(pais_id) | pais_id == "", pais_id := "Brasil"]
# statistics
soildata[, .N, by = .(pais_id)][order(pais_id)]

# estado_id
# Check if the number of unique values in estado_id is 27 (26 states + Federal District)
soildata[estado_id == "", estado_id := NA_character_]
# statistics
soildata[, .N, by = .(estado_id)][order(estado_id)]
# Intersect with Brazilian states
soildata_sf <- unique(soildata[
  is.na(estado_id) & !(is.na(coord_x) | is.na(coord_y)),
  .(dataset_id, observacao_id, coord_x, coord_y)
])
soildata_sf <- sf::st_as_sf(soildata_sf,
  coords = c("coord_x", "coord_y"),
  crs = 4326, remove = FALSE, na.fail = FALSE
)
soildata_sf <- sf::st_transform(soildata_sf, crs = sf::st_crs(brazil))
soildata_sf <- sf::st_join(
  soildata_sf, brazil[, c("abbrev_state")],
  join = sf::st_within, left = TRUE
)
# Extract data.table
soildata_sf <- data.table::as.data.table(soildata_sf)[, .(dataset_id, observacao_id, abbrev_state)]
# Update estado_id in soildata, keeping all of its existing values and adding new ones
soildata <- merge(soildata, soildata_sf, by = c("dataset_id", "observacao_id"), all.x = TRUE)
soildata[is.na(estado_id), estado_id := abbrev_state]
soildata[, abbrev_state := NULL]
# Set manual corrections for estado_id based on dataset_id and observacao_id
soildata[dataset_id == "ctb0053", estado_id := ifelse(is.na(estado_id), "RO", estado_id), by = observacao_id]
soildata[dataset_id == "ctb0711" & observacao_id == "48", estado_id := ifelse(is.na(estado_id), "AM", estado_id)]
soildata[dataset_id == "ctb0831", estado_id := ifelse(is.na(estado_id), "PR", estado_id), by = observacao_id]
# statistics
soildata[, .N, by = .(estado_id)][order(estado_id)]
# dataset_id: ctb0770
# observacao_id: E110 and E84
# soildata[is.na(estado_id), .(dataset_id, observacao_id,  coord_x, coord_y)]

# municipio_id
# Replace "" and "-" with NA
soildata[municipio_id == "", municipio_id := NA_character_]
soildata[municipio_id == "-", municipio_id := NA_character_]
# statistics
soildata[, .N, by = .(municipio_id)]

# Drop column
# organizacao_nome
# drop column
soildata[, organizacao_nome := NA_character_]

# coord_datum
# If coord_datum == "", set to NA
soildata[coord_datum == "", coord_datum := NA_character_]
# If coord_datum is not NA, but coord_x or coord_y are NA, set coord_datum to NA
soildata[!is.na(coord_datum) & ((is.na(coord_x) | is.na(coord_y))), coord_datum := NA_character_]
# If coord_datum == 4326, set to "EPSG:4326"
soildata[coord_datum == "4326", coord_datum := "EPSG:4326"]
# If coord_datum == NA but coord_x and coord_y are not NA, set coord_datum to "EPSG:4326"
soildata[is.na(coord_datum) & !is.na(coord_x) & !is.na(coord_y), coord_datum := "EPSG:4326"]
# statistics
soildata[!(is.na(coord_x) | is.na(coord_y)), .N, by = .(coord_datum)][order(coord_datum)]

# ano_fonte
# If ano_fonte == "", set to NA
soildata[ano_fonte == "", ano_fonte := NA_character_]
# If ano_fonte is NA but data_ano is not NA, set ano_fonte to "original"
soildata[is.na(ano_fonte) & !is.na(data_ano), ano_fonte := "original"]
# statistics
soildata[, .N, by = .(ano_fonte)][order(ano_fonte)]

# data_ano
# Many datasets still have no information for data_ano (sampling year)
soildata[, .N, by = .(data_ano)][order(data_ano)]

# Minimum soil density (dsi)
dsi_min <- 0.5
dsi_min_taxon <- "Organossolo|Húmico|ORGÂNICO|Hístico"
dsi_min_layer <- "O|H|Bw"
# Check if the minimum dsi is greater than or equal to 0.5
min(soildata$dsi, na.rm = TRUE) >= dsi_min
# Check for dsi values less than 0.5 and not in "Organossolo|Húmico" soils or in "O|H" layers
soildata[
  dsi < dsi_min & !grepl(dsi_min_taxon, taxon_sibcs, ignore.case = TRUE) &
  !grepl(dsi_min_layer, camada_nome, ignore.case = TRUE),
  .(taxon_sibcs, camada_nome, dsi, ctc, carbono, dataset_id, observacao_id)
]
# Set those dsi values to NA: we cannot find justification for such low values
soildata[
  dsi < dsi_min & !grepl(dsi_min_taxon, taxon_sibcs, ignore.case = TRUE) &
  !grepl(dsi_min_layer, camada_nome, ignore.case = TRUE),
  dsi := NA_real_
]

# Maximum soil density (dsi)
dsi_max <- 2L
dsi_max_taxon <- "VERTISSOLO|vértico|FLÚVICO"
dsi_max_layer <- "v|t|Cg|Cn"
# Check if the maximum dsi is less than or equal to 2
max(soildata$dsi, na.rm = TRUE) <= dsi_max
# Check for dsi values greater than 2 and not in "VERTISSOLO" soils or in "v|t" layers
soildata[
  dsi > dsi_max & !grepl(dsi_max_taxon, taxon_sibcs, ignore.case = TRUE) &
  !grepl(dsi_max_layer, camada_nome, ignore.case = TRUE),
  .(taxon_sibcs, camada_nome, dsi, areia, carbono, dataset_id, observacao_id)
]
# The values appear to be correct, so we will not change them.

# Maximum pH
ph_max <- 9L
ph_max_taxon <- "carbonático|NÁTRICO|Sódico|SOLODIZADO"
# Check if the maximum pH is less than or equal to 9
max(soildata$ph, na.rm = TRUE) <= ph_max
# Check for pH values greater than 9 and not in "carbonático|NÁTRICO|Sódico|SOLODIZADO" soils
soildata[
  ph > ph_max & !grepl(ph_max_taxon, taxon_sibcs, ignore.case = TRUE),
  .(taxon_sibcs, camada_nome, ph, ctc, carbono, dataset_id, observacao_id)
]
# Set those pH values to NA: we cannot find justification for such high values
soildata[
  ph > ph_max & !grepl(ph_max_taxon, taxon_sibcs, ignore.case = TRUE),
  ph := NA_real_
]

# Minimum pH
ph_min <- 2L
ph_min_taxon <- "TIOMÓRFICO|Húmico"
# Check if the minimum pH is greater than or equal to 2
min(soildata$ph, na.rm = TRUE) >= ph_min
# Check for pH values less than MIN and not in "Tiomórfico|Húmico" soils
soildata[
  ph < ph_min & !grepl(ph_min_taxon, taxon_sibcs, ignore.case = TRUE),
  .(taxon_sibcs, camada_nome, ph, ctc, carbono, dataset_id, observacao_id)
]
# Set those pH values to NA: we cannot find justification for such low values
soildata[
  ph < ph_min & !grepl(ph_min_taxon, taxon_sibcs, ignore.case = TRUE),
  ph := NA_real_
]

# Sum of terrafina and esqueleto == 1000 g/kg
all((soildata$terrafina + soildata$esqueleto) == 1000, na.rm = TRUE)

# Sum clay, silt and sand
# Check if the sum is 1000 g/kg
all((soildata$argila + soildata$silte + soildata$areia) == 1000, na.rm = TRUE)

# Minimum ctc
ctc_min <- 1L
ctc_min_taxon <- "FLÚVICO|arênico|Álic|Distrófico|ESPODOS|Aluvia|fraco|Quartzosas|PODZOL|Latos"
ctc_min_layer <- "C|E|A2"
# Check if the minimum ctc is greater than or equal to MIN
min(soildata$ctc, na.rm = TRUE) >= ctc_min
# Check for pH values less than ctc_min and not in ctc_min_taxon soils and ctc_min_layer layers
# and areia < 500 and ph > 5
soildata[
  ctc < ctc_min & !grepl(ctc_min_taxon, soildata$taxon_sibcs, ignore.case = TRUE) &
  !grepl(ctc_min_layer, soildata$camada_nome, ignore.case = FALSE) & areia < 500 & ph > 5,
  .(taxon_sibcs, camada_nome, ctc, ph, carbono, areia, dataset_id, observacao_id)
]
# Set those ctc values to NA: we cannot find justification for such low values
soildata[
  ctc < ctc_min & !grepl(ctc_min_taxon, soildata$taxon_sibcs, ignore.case = TRUE) &
  !grepl(ctc_min_layer, soildata$camada_nome, ignore.case = FALSE) & areia < 500 & ph > 5,
  ctc := NA_real_
]

# Maximum ctc
ctc_max <- 100
ctc_max_taxon <- "Organossolo|Húmico|vertiso|ORGÂNICO"
ctc_max_layer <- "O|H|v|t"
# Check if the maximum ctc is less than or equal to MAX
max(soildata$ctc, na.rm = TRUE) <= ctc_max
# Check for ctc values greater than ctc_max and not in ctc_max_taxon soils and ctc_max_layer layers
soildata[
  ctc > ctc_max & !grepl(ctc_max_taxon, soildata$taxon_sibcs, ignore.case = TRUE) &
  !grepl(ctc_max_layer, soildata$camada_nome, ignore.case = TRUE) & ph < 5,
  .(taxon_sibcs, camada_nome, ctc, ph, carbono, argila, dataset_id, observacao_id)
]
# Manual corrections
soildata[
  dataset_id == "ctb0718" & observacao_id == "55" & camada_nome == "A3",
  ctc := ifelse(ctc == 140.4, 144.4, ctc)
]
# We will not change other values, as they appear to be correct.

# Export
# Save data in TXT format, saving in "res/tab/"
summary_soildata(soildata)
# Layers: 57077
# Events: 16824
# Georeferenced events: 14334
# Datasets: 255
file_path <- "res/tab/brazilian-soil-dataset-2024.txt"
data.table::fwrite(soildata, file = file_path, sep = "\t", dec = ".", row.names = FALSE, na = "NA")
# Load to SoilData repository!
